stock-monitor/requirements.txt
akshare==1.10.8
pandas==2.0.3
numpy<2
pydantic-settings==2.0.3
python-telegram-bot==20.5
apscheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.9.1
requests>=2.31.0
matplotlib>=3.7.0

# 美股/港股数据支持
yfinance>=0.2.28

# Test dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
